refactor: 모임 채팅방 조회 로직을 별도 서비스로 분리
